+++ src/App.tsx
@@ -984,7 +984,7 @@ function App() {
         data-tauri-drag-region 
         onMouseDown={handleDragStart}
       />
-      
+
       {/* 左侧悬浮导航 */}
       <SideNav
         page={page}